Move CreateFuncHash to util
Refactor generator test util to call any function with function signature
